Quick Fixes & Minor changes, Better Player & NPC streaming, World class cleanup & miscellaneous changes

+++ 727 server/network/com/rs/net/decoders/LogicPacket.java
@@ -4,10 +4,10 @@ import com.rs.cache.io.InputStream;
 
 public class LogicPacket {
 
-	private int id;
+	private byte id;
 	byte[] data;
 
-	public LogicPacket(int id, int size, InputStream stream) {
+	public LogicPacket(byte id, int size, InputStream stream) {
 		this.id = id;
 		data = new byte[size];
 		stream.getBytes(data, 0, size);
