+++ 727 server/plugins/main/impl/rsinterface/EquipmentInterfacePlugin.java
@@ -13,7 +13,7 @@ import main.wrapper.RSInterfaceSignature;
 public class EquipmentInterfacePlugin implements RSInterface {
 
 	@Override
-	public void execute(Player player, int interfaceId, int componentId, int packetId, int slotId, int slotId2) throws Exception {
+	public void execute(Player player, int interfaceId, int componentId, int packetId, byte slotId, int slotId2) throws Exception {
 		if (player.getInterfaceManager().containsInventoryInter())
 			return;
 		if (componentId == 6) {
@@ -97,7 +97,7 @@ public class EquipmentInterfacePlugin implements RSInterface {
 		}
 	}
 	
-	public static void sendRemove(Player player, int slotId) {
+	public static void sendRemove(Player player, byte slotId) {
 		if (slotId >= 15)
 			return;
 		player.stopAll(false, false);
