+++ 727 server/plugins/main/impl/rsinterface/OptionsTabInterfacePlugin.java
@@ -9,7 +9,7 @@ import main.wrapper.RSInterfaceSignature;
 public class OptionsTabInterfacePlugin implements RSInterface {
 
 	@Override
-	public void execute(Player player, int interfaceId, int componentId, int packetId, int slotId, int slotId2) throws Exception {
+	public void execute(Player player, int interfaceId, int componentId, int packetId, byte slotId, int slotId2) throws Exception {
 		if (player.getInterfaceManager().containsInventoryInter())
 			return;
 		if (componentId == 22) {
