+++ 727 server/src/com/rs/game/player/actions/Rest.java
@@ -19,7 +19,7 @@ public class Rest extends Action {
 		index = Utils.random(REST_DEFS.length);
 		player.setResting(true);
 		player.setNextAnimation(new Animation(REST_DEFS[index][0]));
-		player.getAppearance().setRenderEmote(REST_DEFS[index][1]);
+		player.getAppearance().setRenderEmote((short) REST_DEFS[index][1]);
 		return true;
 	}
 
@@ -46,7 +46,7 @@ public class Rest extends Action {
 		player.setResting(false);
 		player.setNextAnimation(new Animation(REST_DEFS[index][2]));
 		player.getEmotesManager().setNextEmoteEnd();
-		player.getAppearance().setRenderEmote(-1);
+		player.getAppearance().setRenderEmote((short) -1);
 	}
 
 }
